pyproject.toml
[tool.poetry]
name = "quant-mvp"
version = "0.1.0"
description = "Interview Prep"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^6.0.3"
django-ninja = "^1.6.2"
polars = "^1.39.3"
pyarrow = "^23.0.1"
httpx = "^0.28.1"
pydantic = "^2.12.5"
pytest = "^9.0.2"
ruff = "^0.15.8"
opentelemetry-api = "^1.40.0"
opentelemetry-sdk = "^1.40.0"
uvicorn = "^0.42.0"
opentelemetry-exporter-otlp = "^1.40.0"
opentelemetry-instrumentation-django = "^0.61b0"
opentelemetry-instrumentation-httpx = "^0.61b0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
